refactoring- added function for hidden and styling changes
