+++ Activity.js
@@ -1,27 +1,3 @@
-// // module.exports = Activity;
-// class Activity {
-//     constructor(category, description, minutes, seconds) {
-//         this.category = category
-//         this.description = description
-//         this.minutes = minutes
-//         this.seconds = seconds
-//         this.completed = false
-//         this.id = null
-//         console.log('This is working')
-//     }
-//     countdown() {
-
-//     }
-
-//     markComplete() {
-
-//     }
-
-//     saveToStorage() {
-
-//     }
-// }
-
 class Activity {
     constructor(category, description, minutes, seconds) {
         this.category = category
@@ -44,5 +20,3 @@ class Activity {
     localStorage.setItem(objectToStore, JSON.stringify(this))
     }
 }
-
-// default export Activity;
