+++ datamodel.js
@@ -1,1 +1,47 @@
 var cards = [];
+
+const compliments = ["Respects",
+"Fantastic",
+"Hats off!",
+"Sensational",
+"Well done",
+"Good job!",
+"You rock!",
+"Nice going",
+"Good show!",
+"Good going!",
+"Good for you!",
+"Good on you!",
+"Good one mate!",
+"I am impressed",
+"Way to go",
+"You did it!",
+"You’re a genius!",
+"You’re the best",
+"You’ve got it",
+"Pat on the back",
+"That’s the way",
+"That’s the best ever",
+"You did it that time!",
+"You make it look easy",
+"You really deserved it",
+"You did that very well",
+"What a good try",
+"Many happy returns",
+"That is better than ever",
+"You’re doing a good job",
+"You’re doing beautifully",
+"You’re really growing up",
+"That’s the way to do it",
+"You are very good at that",
+"You’re on the right track now",
+"You’re really going to town",
+"Keep up the good work",
+"That’s the right way to do it",
+"You have done a great job!",
+"You’re getting better every day",
+"You’ve just about mastered that",
+"Nothing can stop you now",
+"That’s the best you’ve ever done",
+"Keep working, you’re getting better",
+"I’m happy to see you working like that."]
