+++ main.js
@@ -1,4 +1,3 @@
-// var Activity = require('./Activity.js
 
 var studyBtn = document.getElementById('studyBtn')
 var meditateBtn = document.getElementById('meditateBtn')
@@ -26,78 +25,8 @@ var timerWrapper = document.getElementById('timerWrapper')
 var createNewBtn = document.getElementById('createNewBtn')
 var deck = document.getElementById('deck')
 
-// Activity Class Constructor
-// class Activity {
-//     constructor(category, description, minutes, seconds) {
-//         this.category = category
-//         this.description = description
-//         this.minutes = minutes
-//         this.seconds = seconds
-//         this.completed = false
-//         this.id = Date.now()
-//     }
-//     countdown() {
-
-//     }
-
-//     markComplete() {
-
-//     }
-
-//     saveToStorage() {
-//     var objectToStore = this.id.toString();
-//     localStorage.setItem(objectToStore, JSON.stringify(this))
-//     }
-// }
 
 // global
-const compliments = ["Respects",
-"Fantastic",
-"Hats off!",
-"Sensational",
-"Well done",
-"Good job!",
-"You rock!",
-"Nice going",
-"Good show!",
-"Good going!",
-"Good for you!",
-"Good on you!",
-"Good one mate!",
-"I am impressed",
-"Way to go",
-"You did it!",
-"You’re a genius!",
-"You’re the best",
-"You’ve got it",
-"Pat on the back",
-"That’s the way",
-"That’s the best ever",
-"You did it that time!",
-"You make it look easy",
-"You really deserved it",
-"You did that very well",
-"What a good try",
-"Many happy returns",
-"That is better than ever",
-"You’re doing a good job",
-"You’re doing beautifully",
-"You’re really growing up",
-"That’s the way to do it",
-"You are very good at that",
-"You’re on the right track now",
-"You’re really going to town",
-"Keep up the good work",
-"That’s the right way to do it",
-"You have done a great job!",
-"You’re getting better every day",
-"You’ve just about mastered that",
-"Nothing can stop you now",
-"That’s the best you’ve ever done",
-"Keep working, you’re getting better",
-"I’m happy to see you working like that."]
-
-
 startBtn.addEventListener('click', submit)
 startTimerBtn.addEventListener('click', startTimer)
 logActBtn.addEventListener('click', populate)
@@ -109,15 +38,23 @@ window.addEventListener('load', loadStorage)
 //functions
 
 function switchForm() {
-  timerPage.classList.remove('hidden')
-  newActivityPage.classList.remove('hidden')
+  removeHidden(timerPage)
+  addHidden(newActivityPage)
+}
+
+function removeHidden(elementVar) {
+  elementVar.classList.remove('hidden')
+}
+
+function addHidden(elementVar) {
+  elementVar.classList.add('hidden')
 }
 
 function loadStorage() {
   var justIds = Object.keys(localStorage)
   if (justIds.length > 0) {
-    noAct.classList.add('hidden')
-    deck.classList.remove('hidden')
+    addHidden(noAct)
+    removeHidden(deck)
     for (var i = 0; i < justIds.length; i++) {
       var hold = JSON.parse(localStorage.getItem(justIds[i]))
       cards.push(hold)
@@ -129,17 +66,17 @@ function loadStorage() {
 function backtoForm() {
   startTimerBtn.disabled = false
   startTimerBtn.innerText = "START"
-  completedPage.classList.add('hidden')
-  newActivityPage.classList.remove('hidden')
+  addHidden(completedPage)
+  removeHidden(newActivityPage)
   userActivity.value = '';
   userMinutes.value = '';
   userSeconds.value = '';
   for (var i = 0; i < radioBtns.length; i++) {
-      if (radioBtns[i].checked) {
-        radioBtns[i].checked = false;
-      }
+    if (radioBtns[i].checked) {
+      radioBtns[i].checked = false;
     }
   }
+}
 
 
 function inputUserValues() {
@@ -151,12 +88,6 @@ function inputUserValues() {
   }
   displayTimerFunction.innerText = `${userMinutes.value} : ${userSeconds.value}`
   timerEvent.innerText = userActivity.value
- }
-
-
-function switchForm() {
-  timerPage.classList.remove('hidden')
-  newActivityPage.classList.add('hidden')
 }
 
 function preventTimerErrors(event) {
@@ -168,13 +99,12 @@ function preventTimerErrors(event) {
 function submit() {
     event.preventDefault()
     if (!userActivity.value || !userMinutes.value || !userSeconds.value || !isBtnChecked()) {
-        errorWrapper.classList.remove('hidden')
+        removeHidden(errorWrapper)
     } else if ((userMinutes.value || userSeconds.value) == 'e') {
-        errorWrapper.classList.remove('hidden')
+        removeHidden(errorWrapper)
     } else {
         var activityCard = new Activity (whichBtn(), userActivity.value, userMinutes.value, userSeconds.value)
         cards.push(activityCard)
-        // ids.push(activityCard.id)
         activityCard.saveToStorage()
         inputUserValues()
         switchForm()
@@ -192,13 +122,13 @@ function isBtnChecked() {
 }
 
 function whichBtn() {
-    var theRightBtn = null
-    for (var i = 0; i < radioBtns.length; i++) {
-        if (radioBtns[i].checked) {
-        theRightBtn = radioBtns[i].value
-        return theRightBtn
-        }
+  var theRightBtn = null
+  for (var i = 0; i < radioBtns.length; i++) {
+    if (radioBtns[i].checked) {
+      theRightBtn = radioBtns[i].value
+      return theRightBtn
     }
+  }
 }
 
 function setStartBtnColor() {
@@ -224,33 +154,32 @@ function markerColor(category) {
 }
 
 function populate() {
-  noAct.classList.add('hidden')
-  timerPage.classList.add('hidden')
-  completedPage.classList.remove('hidden')
+  addHidden(noAct)
+  addHidden(timerPage)
+  removeHidden(completedPage)
   reloadCards()
 }
 
-  function reloadCards() {
-  deck.classList.remove('hidden')
+function reloadCards() {
+  removeHidden(deck)
   cardWrapper.innerHTML = ''
   for (var i = 0; i < cards.length; i++) {
-    // Change div to a semantic tag
     cardWrapper.innerHTML += `
-    <div class="card">
-        <div class="card-details">
+    <article class="card">
+      <div class="card-details">
         <div>
           <h4>${cards[i].category}</h4>
           <p class="card-time" id="cardTime">${cards[i].minutes} MINS ${cards[i].seconds} SECONDS</p>
         </div>
           <p class="card-activity" id="card-activity">${cards[i].description}</p>
-        </div>
-        <div class="marker-wrapper">
-          <div class="card-marker" style="background: ${markerColor(cards[i].category)}"></div>
-        </div>
       </div>
+      <div class="marker-wrapper">
+        <div class="card-marker" style="background: ${markerColor(cards[i].category)}"></div>
+      </div>
+    </article>
     `
   }
- }
+}
 
 function startTimer() {
   startTimerBtn.disabled = true
@@ -262,28 +191,21 @@ function startTimer() {
     var intervalTimer = setInterval(timeLeft, 1000)
     function timeLeft() {
       totalSeconds--
-      var minutes = Math.floor(totalSeconds / 60);
-      var seconds = Math.floor(totalSeconds % 60);
-        if(seconds < 10 && seconds != '00') {
-          seconds = `0` + `${seconds}`
-          }
-        if(minutes < 10 && seconds != '00') {
-          minutes = `0` + `${minutes}`
-          }
+      var minutes = Math.floor(totalSeconds / 60)
+      var seconds = Math.floor(totalSeconds % 60)
+      if(seconds < 10 && seconds != '00') {
+        seconds = `0` + `${seconds}`
+        }
+      if(minutes < 10 && seconds != '00') {
+        minutes = `0` + `${minutes}`
+        }
       displayTimerFunction.innerText = `${minutes} : ${seconds}`
       if (!totalSeconds) {
         clearInterval(intervalTimer)
         startTimerBtn.innerText = "COMPLETE!"
         displayTimerFunction.innerText = `${compliments[Math.floor(Math.random()*compliments.length)]}`
-        showLogBtn()
-   }
+        removeHidden(logActBtn)
+      }
+    }
   }
- }
 }
-
-function showLogBtn() {
- logActBtn.classList.remove('hidden')
-}
-
-
-// classList add/remove , write a function that takes a parameter
