+++ styles.css
@@ -48,7 +48,7 @@ nav {
 
 h1 {
     margin: 6px;
-    font-size: 28px;
+    font-size: 24px;
 }
 
 /* Main Body */
